+++ src/decayactionsfinder.cc
@@ -73,8 +73,8 @@ ActionList DecayActionsFinder::find_actions_in_cell(
   return actions;
 }
 
-ActionList DecayActionsFinder::find_final_actions(const Particles &search_list,
-                                                  bool /*only_res*/) const {
+ActionList DecayActionsFinder::find_final_actions(
+    const Particles &search_list) const {
   ActionList actions;
   if (find_final_decays_) {
     for (const auto &p : search_list) {